--- src/types/comment.ts
@@ -1,11 +1,0 @@
-export interface Comment {
-  id: number;
-  body: string;
-  postId: number;
-  likes: number;
-  user: {
-    id: number;
-    username: string;
-    fullName: string;
-  };
-}
