--- src/types/index.ts
@@ -1,3 +1,0 @@
-export * from './post';
-export * from './user';
-export * from './comment';
