--- src/types/user.ts
@@ -1,23 +1,0 @@
-export interface User {
-  id: number;
-  username: string;
-  image: string;
-  firstName?: string;
-  lastName?: string;
-  age?: number;
-  email?: string;
-  phone?: string;
-  address?: {
-    address: string;
-    city: string;
-    state: string;
-  };
-  company?: {
-    name: string;
-    title: string;
-  };
-}
-
-export interface UsersResponse {
-  users: User[];
-}
